team backend restructure for circular reference
