+++ darwin/future/data_objects/team.py
@@ -5,6 +5,7 @@ from typing import List, Optional, Tuple
 from pydantic import validator
 
 from darwin.future.core.client import ClientCore
+from darwin.future.core.team.get_raw import get_team_raw
 from darwin.future.data_objects.dataset import DatasetList
 from darwin.future.data_objects.team_member_role import TeamMemberRole
 from darwin.future.data_objects.validators import parse_name
@@ -68,43 +69,5 @@ class TeamCore(DefaultDarwin):
         """
         if not team_slug:
             team_slug = client.config.default_team
-        return get_team(client, team_slug)
-
-
-TeamList = List[TeamCore]
-
-
-def get_team(client: ClientCore, team_slug: Optional[str] = None) -> TeamCore:
-    """Returns the team with the given slug from the client
-
-    Args:
-        client (Client): Core client object
-        team_slug (Optional[str], optional): team slug str, Defaults to None.
-
-    Returns:
-        Team: Team object retrieved from the client with the given slug
-    """
-    if not team_slug:
-        team_slug = client.config.default_team
-    response = client.get(f"/teams/{team_slug}/")
-    return TeamCore.parse_obj(response)
-
-
-def get_team_members(client: ClientCore) -> Tuple[List[TeamMemberCore], List[Exception]]:
-    """Returns a list of team members for the given client
-
-    Args:
-        client (Client): Core client object
-
-    Returns:
-        Tuple[List[TeamMember], List[Exception]]: List of team members and list of errors if any
-    """
-    response = client.get(f"/memberships")
-    members = []
-    errors = []
-    for item in response:
-        try:
-            members.append(TeamMemberCore.parse_obj(item))
-        except Exception as e:
-            errors.append(e)
-    return members, errors
+        url = client.config.api_endpoint + f"teams/{team_slug}"
+        return TeamCore.parse_obj(get_team_raw(client.session, url))
