+++ darwin/future/meta/objects/team_member.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from darwin.future.core.client import ClientCore
-from darwin.future.data_objects.team import TeamMemberCore, get_team_members
+from darwin.future.data_objects.team import TeamMemberCore
 from darwin.future.data_objects.team_member_role import TeamMemberRole
 from darwin.future.meta.objects.base import MetaBase
