+++ darwin/future/meta/queries/team_member.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import List
 
+from darwin.future.core.team.get_team import get_team_members
 from darwin.future.core.types.query import Param, Query, QueryFilter
-from darwin.future.data_objects.team import get_team_members
 from darwin.future.meta.objects.team_member import TeamMember
