+++ darwin/future/tests/meta/queries/test_team_member.py
@@ -4,7 +4,6 @@ import pytest
 import responses
 
 from darwin.future.core.client import ClientCore
-from darwin.future.data_objects.team import TeamMemberCore
 from darwin.future.data_objects.team_member_role import TeamMemberRole
 from darwin.future.meta.objects.team_member import TeamMember
 from darwin.future.meta.queries.team_member import TeamMemberQuery
